pyproject.toml
[project]
name = "waves"
version = "0.1.0"
description = "Simple audio recordings hosting server"
authors = [{ name = "Luca Scotton", email = "luca@lucascott.com" }]
requires-python = "~=3.13"
readme = "README.md"
dependencies = [
    "flask~=3.0",
    "waitress~=3.0",
    "prometheus-flask-exporter~=0.23",
    "pyyaml~=6.0",
    "markdown~=3.10",
    "flask-caching~=2.3",
]

[dependency-groups]
dev = ["ruff"]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/
    "E",
    "W",
    "F",
    "PL",
    "I",
    "N",
    "UP",
]
ignore = [
    # disabling this as it disagrees with the formatter
    # in cases like comments or long docstrings
    "E501",
]
